--- src/main/java/com/example/wegather/global/auth/IAuthenticationManager.java
@@ -1,12 +1,0 @@
-package com.example.wegather.global.auth;
-
-import com.example.wegather.auth.MemberDetails;
-import org.springframework.security.core.Authentication;
-
-public interface IAuthenticationManager {
-  Authentication getAuthentication();
-
-  MemberDetails getPrincipal();
-
-  String getUsername();
-}
