+++ src/main/java/com/example/wegather/group/domain/SmallGroupService.java
@@ -1,7 +1,7 @@
 package com.example.wegather.group.domain;
 
 import com.example.wegather.auth.MemberDetails;
-import com.example.wegather.global.auth.AuthenticationManager;
+import com.example.wegather.global.auth.AuthenticationManagerImpl;
 import com.example.wegather.global.customException.AuthenticationException;
 import com.example.wegather.global.vo.Address;
 import com.example.wegather.group.domain.repotitory.SmallGroupRepository;
@@ -27,7 +27,7 @@ public class SmallGroupService {
   private static final String DO_NOT_HAVE_AUTHORITY_TO_DELETE_GROUP = "소모임을 삭제할 권한이 없습니다.";
   private final SmallGroupRepository groupRepository;
   private final MemberRepository memberRepository;
-  private final AuthenticationManager authManager;
+  private final AuthenticationManagerImpl authManager;
 
   @Transactional
   public SmallGroup addGroup(CreateSmallGroupRequest request, String username) {
